chore: remove useStore in MonacoEditor
